move PanelBuilder logic to routing
