+++ src/javascripts/ng-admin/Main/component/controller/DashboardController.js
@@ -10,11 +10,11 @@ define(function (require) {
      * @param {PanelBuilder} PanelBuilder
      * @constructor
      */
-    function DashboardController($scope, $state, PanelBuilder) {
-        this.$scope = $scope;
+    function DashboardController($scope, $state, collections, entries, hasEntities) {
         this.$state = $state;
-
-        this.retrieveCollectionsAndData(PanelBuilder);
+        this.collections = collections;
+        this.entries = entries;
+        this.hasEntities = hasEntities;
 
         $scope.$on('$destroy', this.destroy.bind(this));
     }
@@ -23,30 +23,11 @@ define(function (require) {
         this.$state.go(this.$state.get('list'), { entity: entityName });
     }
 
-    /**
-     * Retrieve all dashboard panels
-     */
-    DashboardController.prototype.retrieveCollectionsAndData = function (PanelBuilder) {
-        let collections = PanelBuilder.getCollections();
-        this.entries = {};
-
-        var searchParams = this.$state.params;
-        var sortField = 'sortField' in searchParams ? searchParams.sortField : null;
-        var sortDir = 'sortDir' in searchParams ? searchParams.sortDir : null;
-
-        PanelBuilder.getEntries(sortField, sortDir).then(entries => {
-            this.entries = entries;
-            this.collections = collections; // we set collectrions only when entries are fetched to avoid double draw
-        });
-        this.hasEntities = PanelBuilder.hasEntities();
-    };
-
     DashboardController.prototype.destroy = function () {
-        this.$scope = undefined;
         this.$state = undefined;
     };
 
-    DashboardController.$inject = ['$scope', '$state', 'PanelBuilder'];
+    DashboardController.$inject = ['$scope', '$state', 'collections', 'entries', 'hasEntities'];
 
     return DashboardController;
 });
